fix pdf export missing rows
When dividing the table for multi-column layout there was a bug in the algorithm: After computing the amount of rows available it increments the current row index by that count and takes the data from *after* the last end index up to, *excluding*, the new end index.
While the first one was intentional to avoid negative indexing on the first page, excluding the new end index is entirely unnecessary and was likely introduced during a refactor.

fixes #485